Change between modes with props registering
