+++ src/components/pages/Gamemode/Gamemode.js
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 // import history from './history';
 import { connect } from 'react-redux';
 import { tasks } from '../../../state/actions';
@@ -10,46 +10,63 @@ import { render } from 'react-dom';
 
 const Gamemode = ({ ...props }) => {
   const { push, location } = useHistory();
-  // const [sP, setsP] = useState(false);
+  const [sP, setsP] = useState(false);
   //   history.push('/gamemode/single');
-
-  //   useEffect(() => {
-
-  //     ff();
-  //   }, []);
-
-  // const singled = () => {
-  //   if (location.pathname === '/gamemode/single' && sP === true) {
-  //     push('/gamemode');
-  //     setsP(false);
-  //     console.log('nal', this.state.hasRead);
-  //   } else if (location.pathname === '/gamemode' && sP === false) {
-  //     push('/gamemode/single');
-  //     setsP(true);
-  //   }
-  // };
-
-  //   console.log(history);
-  const clicker = () => {
+  const propInit = () => {
     props.child.gamemode = {
-      mode: 'single',
+      mode: 'select',
       read: null,
       write: null,
       draw: null,
     };
-    console.log(props.child);
   };
+
+  useEffect(() => {
+    propInit();
+  }, []);
+
+  const singled = () => {
+    if (location.pathname === '/gamemode/single' && sP === true) {
+      push('/gamemode');
+      props.child.gamemode = {
+        mode: 'select',
+        read: null,
+        write: null,
+        draw: null,
+      };
+      setsP(false);
+      console.log('nal', props.child);
+    } else if (location.pathname === '/gamemode' && sP === false) {
+      push('/gamemode/single');
+      props.child.gamemode = {
+        mode: 'single',
+        read: null,
+        write: null,
+        draw: null,
+      };
+      console.log(props.child);
+      setsP(true);
+    }
+  };
+
+  //   console.log(history);
   return (
-    <div>
-      <button onClick={clicker}>hi</button>
-    </div>
+    (sP && (
+      <Link to="/gamemode">
+        <Route path="/gamemode" component={GamemodeButton} />
+        <button onClick={singled}> Single Player</button>
+      </Link>
+    )) || (
+      <Link to="/gamemode/single">
+        <button onClick={singled}> Single Player</button>
+        <Route path="/gamemode/single" component={GamemodeButton} />
+      </Link>
+    )
   );
 };
 export default connect(
   state => ({
     child: state.child,
-    faceoffs: state.faceoffs,
-    votes: state.votes,
   }),
   {}
 )(Gamemode);
